Moved main menu to a drawer Puzzle is now the main page
