+++ lib/l10n/app_en.arb
@@ -1,5 +1,5 @@
 {
-    "appName":"ISCTE_50_Years",
+    "appName":"ISCTE 50 Years",
     "scanCodeButton": "Scan Code",
     "createCode": "Create Code",
     "scanACode": "Scan a code",
